+++ README.md
@@ -2,8 +2,17 @@ community
 ===
 An approach to developing simplified API(s) as a boilerplate in Golang leveraging MongoDB and JSON-API.
 
-####why?
+#### why?
 I have spent the last sixteen months struggling to develop services using Golang in a way that seems successful.  Having some measureable amount of success or being successful includes having repeatable steps and a foundation for progress.  I have developed, refactored, deleted, and started over more times than I can count and now I am committing myself to going in one direction, forward.
 
-####approach
+#### approach
 Each major commit or branch of this code will represent some lesson or lessons I am trying to learn.  They will likely be formalized into more of an article at some point for me to declare some small amount of victory with this approach, but we shall see.
+
+#### inspiration
+I continually monitor the progression of great code coming out of my friends and colleagues and I envy their ability to write stuff that doesn't suck.  I am absorbing their brans to help my code not suck nearly as bad.
+
+THanks to all of the various [make a restful api](https://thenewstack.io/make-a-restful-json-api-go/) in golang [articles](https://github.com/ant0ine/go-json-rest) that are [all over the interwebs](https://www.nicolasmerouze.com/how-to-render-json-api-golang-mongodb/) for giving me plenty of information to read.
+
+#### useful links
+
+* [json-api](http://jsonapi.org/) - is where I am trying to get the infos for making this not suck entirely.
